# config/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "catalog",
        "version": "1.0",
        "description": "My-Condition-Dataカタログサービス"
    },
    "servers": [
        {
            "url": "/catalog"
        }
    ],
    "paths": {
        "/name": {
            "post": {
                "tags": [
                    "アクティベート"
                ],
                "description": "カタログサービスをアクティベートをします。",
                "requestBody": {
                    "description": "カタログ情報",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "カタログのアクティベートおよび名称変更時共通リクエストボディ",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "カタログ名称",
                                        "example": "PLRカタログ"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "カタログ説明",
                                        "example": "組織が運営するPLR基盤が提供するデータカタログです。"
                                    },
                                    "ext_name": {
                                        "type": "string",
                                        "description": "extのネームスペース",
                                        "example": "test-org"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "カタログのアクティベートに成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "カタログのアクティベートおよび名称変更時共通レスポンス",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "description": "カタログID(UUID)",
                                            "example": "nnnnnnnn-nnnn-4nnn-nnnn-nnnnnnnnnnnn"
                                        },
                                        "name": {
                                            "type": "string",
                                            "description": "カタログ名称",
                                            "example": "PLRカタログ"
                                        },
                                        "description": {
                                            "type": "string",
                                            "description": "カタログ説明",
                                            "example": "組織が運営するPLR基盤が提供するデータカタログです。"
                                        },
                                        "ext_name": {
                                            "type": "string",
                                            "description": "extのネームスペース",
                                            "example": "test-org"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "カタログ名称"
                ],
                "description": "アクティベート済のカタログ名称を取得します。",
                "responses": {
                    "200": {
                        "description": "カタログ情報の取得に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "カタログのアクティベートおよび名称変更時共通レスポンス",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "description": "カタログID(UUID)",
                                            "example": "nnnnnnnn-nnnn-4nnn-nnnn-nnnnnnnnnnnn"
                                        },
                                        "name": {
                                            "type": "string",
                                            "description": "カタログ名称",
                                            "example": "PLRカタログ"
                                        },
                                        "description": {
                                            "type": "string",
                                            "description": "カタログ説明",
                                            "example": "組織が運営するPLR基盤が提供するデータカタログです。"
                                        },
                                        "ext_name": {
                                            "type": "string",
                                            "description": "extのネームスペース",
                                            "example": "test-org"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "カタログ名称"
                ],
                "description": "アクティベート済のカタログ名称変更を行います。",
                "requestBody": {
                    "description": "カタログ情報",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "カタログのアクティベートおよび名称変更時共通リクエストボディ",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "description": "カタログ名称",
                                        "example": "PLRカタログ"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "カタログ説明",
                                        "example": "組織が運営するPLR基盤が提供するデータカタログです。"
                                    },
                                    "ext_name": {
                                        "type": "string",
                                        "description": "extのネームスペース",
                                        "example": "test-org"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "カタログ情報の更新に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "カタログのアクティベートおよび名称変更時共通レスポンス",
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "description": "カタログID(UUID)",
                                            "example": "nnnnnnnn-nnnn-4nnn-nnnn-nnnnnnnnnnnn"
                                        },
                                        "name": {
                                            "type": "string",
                                            "description": "カタログ名称",
                                            "example": "PLRカタログ"
                                        },
                                        "description": {
                                            "type": "string",
                                            "description": "カタログ説明",
                                            "example": "組織が運営するPLR基盤が提供するデータカタログです。"
                                        },
                                        "ext_name": {
                                            "type": "string",
                                            "description": "extのネームスペース",
                                            "example": "test-org"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/model": {
            "post": {
                "tags": [
                    "追加"
                ],
                "description": "カタログ項目(モデル)を追加します",
                "requestBody": {
                    "description": "追加したいカタログ項目(モデル)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/addModelRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "カタログ項目(モデル)の追加に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/catalogResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/built_in": {
            "post": {
                "tags": [
                    "追加"
                ],
                "description": "カタログ項目(組み込み)を追加します",
                "requestBody": {
                    "description": "追加したいカタログ項目(組み込み)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "カタログ項目(組み込み)登録のリクエストボディ",
                                "properties": {
                                    "catalogItem": {
                                        "$ref": "#/components/schemas/catalogItem"
                                    },
                                    "template": {
                                        "type": "object",
                                        "description": "項目テンプレート",
                                        "properties": {
                                            "prop": {
                                                "type": "array",
                                                "nullable": true,
                                                "description": "項目テンプレート",
                                                "items": {
                                                    "$ref": "#/components/schemas/propItems"
                                                },
                                                "example": [
                                                    {
                                                        "key": "sample1",
                                                        "type": "string"
                                                    },
                                                    {
                                                        "key": "sample2",
                                                        "type": {
                                                            "of": "item",
                                                            "_code": {
                                                                "value": -9,
                                                                "ver": 1
                                                            },
                                                            "candidate": null
                                                        }
                                                    },
                                                    {
                                                        "key": "sample3",
                                                        "type": {
                                                            "of": "inner",
                                                            "inner": "Sample3"
                                                        }
                                                    }
                                                ]
                                            },
                                            "value": {
                                                "type": "array",
                                                "nullable": true,
                                                "description": "プロパティ値の代入定義",
                                                "items": {
                                                    "type": "object",
                                                    "properties": {
                                                        "key": {
                                                            "type": "string",
                                                            "description": "キー名"
                                                        },
                                                        "value": {
                                                            "$ref": "#/components/schemas/tempValues",
                                                            "description": "設定する属性値"
                                                        }
                                                    }
                                                },
                                                "example": [
                                                    {
                                                        "key": "index",
                                                        "value": "1.2"
                                                    },
                                                    {
                                                        "key": "format",
                                                        "value": {
                                                            "_code": {
                                                                "value": 99,
                                                                "ver": 1
                                                            }
                                                        }
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "inner": {
                                        "type": "array",
                                        "description": "プロパティ定義",
                                        "items": {
                                            "$ref": "#/components/schemas/innerItems"
                                        },
                                        "example": [
                                            {
                                                "name": "Sample3",
                                                "itemTemplate": {
                                                    "inherit": null,
                                                    "prop": [
                                                        {
                                                            "key": "sample3_1",
                                                            "type": "string"
                                                        },
                                                        {
                                                            "key": "sample3_2",
                                                            "type": "number"
                                                        }
                                                    ]
                                                },
                                                "inner": null
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "カタログ項目(組み込み)の追加に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/catalogResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/ns": {
            "get": {
                "tags": [
                    "ネームスペース取得"
                ],
                "description": "ネームスペースを指定したコードの前方一致で取得します",
                "parameters": [
                    {
                        "name": "ns",
                        "description": "ネームスペース",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "example": "catalog/model/unit/si"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "カタログ項目の取得に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "ネームスペース取得レスポンス",
                                    "properties": {
                                        "ns": {
                                            "type": "array",
                                            "description": "ネームスペースの共通リクエストボディ配列",
                                            "items": {
                                                "$ref": "#/components/schemas/nsResponse"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ns/{nsId}": {
            "get": {
                "tags": [
                    "ネームスペース取得"
                ],
                "description": "ネームスペースをネームスペースIDを指定して取得します",
                "parameters": [
                    {
                        "name": "nsId",
                        "description": "ネームスペースID",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ネームスペースの取得に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/nsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ns/model": {
            "post": {
                "tags": [
                    "ネームスペース追加"
                ],
                "description": "カタログ項目(モデル)用ネームスペースを追加します",
                "requestBody": {
                    "description": "追加したいカタログ項目(モデル)用ネームスペース",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/nsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "カタログ項目(モデル)用ネームスペースの追加に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/nsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ns/model/{nsId}": {
            "put": {
                "tags": [
                    "ネームスペース更新"
                ],
                "description": "カタログ項目(モデル)用ネームスペースを更新します",
                "parameters": [
                    {
                        "name": "nsId",
                        "description": "更新したいカタログ項目(モデル)用ネームスペースID",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "description": "更新したいカタログ項目(モデル)用ネームスペース",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/nsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "カタログ項目(モデル)用ネームスペースの更新に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/nsResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ネームスペース削除"
                ],
                "description": "カタログ項目(モデル)用ネームスペースを削除します",
                "parameters": [
                    {
                        "name": "nsId",
                        "description": "削除したいカタログ項目(モデル)用ネームスペースID",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "削除に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "description": "削除のリクエストボディ",
                                    "properties": {
                                        "nsId": {
                                            "type": "number",
                                            "example": 1
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/ns/built_in": {
            "post": {
                "tags": [
                    "ネームスペース追加"
                ],
                "description": "カタログ項目(組み込み)用ネームスペースを追加します",
                "requestBody": {
                    "description": "追加したいカタログ項目(組み込み)用ネームスペース",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/nsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "カタログ項目(組み込み)用ネームスペースの追加に成功した際のレスポンスコード",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/nsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "catalogResponse": {
                "type": "object",
                "description": "カタログ項目",
                "properties": {
                    "catalogItem": {
                        "$ref": "#/components/schemas/catalogItem"
                    },
                    "template": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true,
                        "description": "項目テンプレート",
                        "properties": {
                            "_code": {
                                "$ref": "#/components/schemas/_code"
                            }
                        },
                        "example": {
                            "_code": {
                                "_value": 1,
                                "_ver": 1
                            }
                        }
                    },
                    "prop": {
                        "type": "array",
                        "nullable": true,
                        "description": "項目テンプレート",
                        "items": {
                            "$ref": "#/components/schemas/propItems"
                        },
                        "example": null
                    }
                }
            },
...
            "propItems": {
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/propItem1"
                    },
                    {
                        "$ref": "#/components/schemas/propItem2"
                    },
                    {
                        "$ref": "#/components/schemas/propItem3"
                    },
                    {
                        "$ref": "#/components/schemas/propItem4"
                    }
                ]
            },
...
        }
    }
}